Added support for translating inputs & outputs as the pawan.krd model is bad at languages other than english
Add "translate":"sv" to a preset, to make it auto translate the users messages to english, and the responses back to the specified language. This way, everything is english in server side, and only translated to the specified language when sending the response to discord. "sv" can be replaced with a language code.