Added support for non fixed botnames
Non fixed botnames in a preset allow the bot to choose which character to speak as for each response, allowing support for multiple characters.
A botname should still be specified, as it is used in case the bot ever tries to speak as a user, it defaults to speaking as the botname.